Revert "chore: Export `ResultsFilter` and associated types from `src/types/index.ts` to enable client & server-side importing"
This reverts commit fc8ad58d0e75e9bd1605bfeaa7e5dec68fe640b7.
