+++ src/app/src/pages/eval/components/ResultsFilters/FiltersForm.test.tsx
@@ -1,11 +1,10 @@
-import { Box } from '@mui/material';
-import { createTheme, ThemeProvider } from '@mui/material/styles';
-import { fireEvent, render, screen } from '@testing-library/react';
+import { ThemeProvider, createTheme } from '@mui/material/styles';
+import { render, screen, fireEvent } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
-import { useTableStore } from '../store';
+import { vi, describe, it, expect, beforeEach } from 'vitest';
 import FiltersForm from './FiltersForm';
-import type { ResultsFilter } from '@promptfoo/types';
+import { useTableStore, type ResultsFilter } from '../store';
+import { Box } from '@mui/material';
 
 vi.mock('../store', () => ({
   useTableStore: vi.fn(),
