+++ src/app/src/pages/eval/components/store.test.ts
@@ -3,8 +3,8 @@ import { Severity } from '@promptfoo/redteam/constants';
 import { act } from '@testing-library/react';
 import { v4 as uuidv4 } from 'uuid';
 import { beforeEach, describe, expect, it, type Mock, vi } from 'vitest';
-import { useTableStore } from './store';
-import type { EvaluateTable, PromptMetrics, ResultsFile, ResultsFilter } from '@promptfoo/types';
+import { type ResultsFilter, useTableStore } from './store';
+import type { EvaluateTable, PromptMetrics, ResultsFile } from '@promptfoo/types';
 
 vi.mock('uuid', () => ({
   v4: vi.fn(),
